LPS-181233 Move CTSQLModeThreadLocal into portal-kernel so that we can set it from portal-kernel code
